INC-1165 Case Note by type performance improvements

+++ src/main/java/uk/gov/justice/hmpps/prison/api/model/CaseNoteUsageByBookingIdDto.kt
@@ -2,7 +2,7 @@ package uk.gov.justice.hmpps.prison.api.model
 import java.time.LocalDateTime
 
 data class CaseNoteUsageByBookingIdDto(
-  val bookingId: Int?,
+  val bookingId: Long?,
   val caseNoteType: String?,
   val caseNoteSubType: String?,
   val numCaseNotes: Int?,
